Split file into 2 screens and the counter function
